[core] introduce reactive Signal (#1082)
This new effect is in honor of the 600th person to start the repository.
Congrats and thank you @diegobernardes!! 🎉

### Problem

Kyo doesn't offer a solution to handle reactive signals, which is a
common need in UI solutions for example.

### Solution

Introduce `Signal` to `kyo-core` with mutable and constant versions and
methods to stream the state.

### Notes

- This is inspired by fs2's `Signal`/`SignallingRef` but the
implementation is more optimized with two separate `AtomicRef`s to
manage the state and listeners, which reduces allocations.
- Please see the scaladocs for an explanation of the functionality and
implementation.
- I had to add a method to allow masking unsafe promises.
- I've also changed the atomic classes to provide `updateAndGet` +
`getAndUpdate` for consistency.

---------

Signed-off-by: Akhilender Bongirwar <akhilenderb9@gmail.com>
Co-authored-by: Adam Hearn <22334119+hearnadam@users.noreply.github.com>
Co-authored-by: Matteo Bilardi <matteobilardi@outlook.com>
Co-authored-by: Akhilender Bongirwar <112749383+akhilender-bongirwar@users.noreply.github.com>