[offheap] disable scaladoc generation to fix build (#1098)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

### Problem

The `release` workflow in main is failing because scaladoc generation of
the `kyo-offheapNative` project fails with a compiler crash.

### Solution

I've tried to fix the issue again without success so, as a workaround,
this PR disables scaladoc generation for the module but only for Native.
Given that we direct users to the JVM scaladocs, it shouldn't be an
issue.