[prelude] introduce Env.getAll/useAll + scaladoc improvements (#1099)
### Problem

The `Env` effect doesn't provide a way to obtain multiple values at
once. Also, the scaladocs of `Env` and `Local` don't provide a good
mental model for when to use the effects since they're similar

### Solution

Introduce `Env.getAll` and `Env.useAll` that allows the user to request
multiple values via a type intersection, operating at the level of
`TypeMap`. Additionally, improve the scaladocs of `Env` and `Local` to
provide a better mental model regarding the effects.

### Notes

- I've renamed `runTypeMap` to `runAll` for consistency