[core] preserve Async.zip concurrency + support up to 10 computations (#1101)
### Problem

The new `Async.zip` methods don't preserve the concurrency of the number
of computations being zipped. I think that's not an ideal behavior since
the goal of these APIs is to execute computations concurrently.

### Solution

Ensure the concurrency is preserved by specifying it when calling
`collectAll`.

### Notes

I've also added methods for up to 10 computations.