remove Ops classes (#1114)
### Problem

We used to use `Ops` classes to encode methods with multiple type
parameter groups. The latest versions of the compiler include the new
clause interleaving feature that provides a similar mechanism. I was
initially reluctant to adopt it because I thought it could make it
harder for people reading the codebase to understand the method
signatures but there are important drawbacks with the current `Ops`
classes:

- The scaladoc generation isn't great because there's the op builder
method indirection
- Use of `pipe` with methods using `Ops` classes require an explicit
function. For instance, `a.pipe(Abort.run)` doesn't compile but
`a.pipe(Abort.run(_))` does.
- The noise of the `Ops` classes makes the code less readable

### Solution

Migrate the all `Ops` classes to clause interleaving. To have some
consistency, I used the `Frame` as the implicit to separate type
parameter groups since the compiler currently requires parameters
between type parameter groups.

### Notes

- There are no changes to the method bodies nor scaladocs.