[kernel] rename `pipe` to `handle` (#1115)
### Problem

The `pipe` method is convenient for effect handling but I think using it
for arbitrary transformations isn't very intuitive.

### Solution

Rename `pipe` to `handle` to more clearly indicate that the method is
meant to be used for effect handling. It still supports arbitrary
transformations but the scaladocs are also updated to indicate that it's
meant primarily for effect handling.