[core] fix interrupts leak (#1125)
Fixes #1122

### Problem

When a parent `Fiber` is executing and encounters a new child `Fiber` to
be waited on, the parent needs to register a callback to propagate
interrupts to the child in case there's an interrupt during the waiting.
The issue is that, once the child fiber finishes and the parent resumes,
the interrupt callback for the child isn't necessary anymore but it
stays in the pending list until the parent fully completes execution.
This leads to a memory leak when a parent fiber handles many children
during its execution.

### Solution

When the parent fiber resumes execution, remove the interrupt callback
for the completed child.

### Notes

I've also included a fast-path optimization: if the child fiber is
already completed, the parent resumes execution immeditely without
suspending. There's a performance penalty with this change due to the
removal of a few `inline`s in `IOPromise` that should be easily offset
with this additional optimization.