mark classes as Serializable (#1132)
### Problem

We currently don't mark classes as `Serializable`, which prevents java
serialiazation.

### Solution

I searched the codebase for class declarations to introduce the ones
that were missing. I made an exception for `Fiber`/`IOTask` and classes
using them since serializing the state of a running fiber would be
problematic. Classes that hold fibers like `Meter` are also not marked.

### Notes

- Case classes are automatically marked as `Serializable`.
- `Safepoint` requires custom serialization to restore from the thread
local