[playwright] introduce Browser integration (#1113)
### Problem

Kyo doesn't offer an effect to interact with browsers, which is a common
feature in AI solutions these days.

### Solution

Introduce an integration with [Playwright](https://playwright.dev/) that
provides programmatic browser access. The effect has two APIs: the
regular methods in the `Browser` companion and the `Op` classes, which
are convenient for LLM access via tools.