Stream merge (#1123)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

Addresses #679 

### Problem
<!--
Explain here the context, and why you're making this change. What is the
problem you're trying to solve?
-->

We need to be able to merge concurrent streams. This is probably the
last crucial stream feature that's missing.

### Solution
<!--
Describe your solution. Focus on helping reviewers understand your
technical approach and implementation decisions.
-->

- My approach is fairly naive: create a channel, run source streams into
it in concurrent background fibers, and stream from the channel.
- The trickiest part of this is to signal when the stream should
complete. The simplest way to do so is to wrap the streamed elements in
`Maybe` and signal an end to the stream by putting `Absent` in the
channel. Since there is no `Tag` for `Maybe`, however, I am using
`Option` which is not ideal for performance.
- Merging requires `Async`, which is not in `kyo-prelude`, so I also had
to work out how to provide an async streaming API without breaking the
project structure. To do this, I've added extension methods in
`kyo-core` both on the `Stream` companion object (factory methods) and
on `Stream[V, S]` instances (combinators).
- The API I've provided allows you to merge two or more streams giving
you the option of whether to stop streaming when all streams finish,
when any of them finishes, or (when merging only two streams) when the
first or second stream finishes. Unlike ZIO, I haven't made
`HaltingStrategy` a parameter of any method, since I never used this,
but just made different methods for the different behaviors.

### Notes
<!--
Add any important additional information as bullet points, such as:
- Implementation details reviewers should know about
- Open questions and concerns
- Limitations
-->

- Since unions are unsupported, this PR only supports merging streams of
the same exact type.
- To help with the implementation, I added `collect` and `collectWhile`
methods to `Stream`, which I think are valuable additions to the API
independent of their utility to this PR

---------

Co-authored-by: Flavio Brasil <fwbrasil@users.noreply.github.com>