Resource: avoid leaking `Queue` based Finalizer details (#1137)
### Problem
`Finalizer` could be implemented a number of ways. Exposing the `Queue`
makes the implementation hard to change. Additionally, this makes it
more complex when we support hierarchical scopes in #1131.

### Solution
Use an `abstract class`, exposing only the necessary behaviors. In a
followup, I will work towards a rename to `Scope` (optional?), then
adding support for `Scope.fork` or some similar naming.

---------

Co-authored-by: Flavio Brasil <fwbrasil@users.noreply.github.com>