[monix] remove integration (#1147)
### Problem

We've been having build failures related to interrupts in the Monix
integration for some time now. I've looked at the failures a few times
and believe it's an issue on Monix's side.

### Solution

Given that the last release is from 2022, it doesn't seem worth trying
to fix it so this PR removes the integration instead.