[data] introduce optimized override for Chunk.apply (#1161)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

### Problem

We recently introduced the `Single` version of `Chunk`. It's under the
`private[kyo]` internal object so it shouldn't leak to users. Also, we'd
need to migrate all single-elements calls to `Chunk.apply` to
`Chunk.single` to leverage the optimization, which isn't ideal.

### Solution

Don't expose `Single` in user-facing APIs and override `Chunk.apply` to
create `Single` internally in case a single element is passed.