[stm] retry Abort when transaction is inconsistent (#1169)
Fixes #1158

### Problem

When a transaction is rolled back due to an `Abort`, the `STM` effect
simply propagates the error. This can be a surprising behavior when the
`Abort` happened due to an inconsistent state during the transaction
execution, which is an expected scenario given that consistency is
checked at commit time.

### Solution

In case of errors, check if the ref log is still valid and, if not, wrap
the error into `FailedTransaction` to trigger a retry in case the
schedule permits.

### Notes

- I've added `Abort.recoverError` and `Abort.foldError` since it seems a
relatively common pattern.

---------

Co-authored-by: Adam Hearn <22334119+hearnadam@users.noreply.github.com>