fix: NPE when enriching exceptions (#1174)
### Problem
When enriching exceptions, null frames could cause a NPE. See bug #1172 

### Solution
Check for null frames when initially pushed, as well as when building
the saved trace.

### Notes
There is probably an issue with `IOTask` pushing null traces, but I
wasn't able to minimize it quickly. This should be slightly more robust
as even if users accidentally push null frames somehow there is
protection.