[data] new Tag with sub-type check caching (#1171)
Fixes #551, #552

### Problem

The current `Tag` implementation has a few major limitations, which is
the main reason we haven't made user-defined effects a first-class
feature yet.

- Sub-type checking based on the tag string is too expensive. It's
designed to avoid allocations but the logic itself is too heavyweight.
Because of that, `ArrowEffect` and `ContextEffect` use `=:=` when
matching effect suspensions to their handlers. Effects that have type
parameters with variance currently need to rely erased tags due to that
(see `Env`, `Abort`, and the lack of variance in `Emit`).
- The separate APIs for intersections and union types is clunky and
limits their use.
- Tags can't be composed dynamically at run time, unlike `SafeClassTag`,
which offers the `|` and `&` methods.
- The implementation currently ignores type parameters in super classes,
which causes #551 and #552.
- Literal types are unsupported, which is problematic since we have
`Abort.literal` now.

### Solution

We've attempted to rewrite `Tag` following the current approach with all
information in the string a few times already without success. The
latest attempt was by @hearnadam in #948, which is the origin of
`TagTestMacro` in this PR. I think we could eventually get to an
implementation with the properties we need but the high complexity of
the code and the constraints regarding allocations would likely become a
maintenance burden in the long run.

This PR adopts a new approach to encode tags by representing them as
regular case classes in the new `Tag.Type` hierarchy. When inferring
tags, the macro reflects on the type and generates a compile-time
constant string as before but the payload is the serialized `Type`
object. If the tag can't be fully materialized statically because it
depends on other runtime-provided tags, the macro then falls back to
allocating an actual `Type` instance.

In addition, a highly-optimized caching mechanism is in place to provide
low-overhead sub-type checking. It has two layers: a cache for the check
itself and, in case of misses in this initial layer, another cache is
used to ensure tags are parsed only once. Please see the scaladocs for
more information and reasoning since it isn't a very orthodox approach.

I've also added a restriction: tags in the `kyo` package must be static
to avoid allocations. For example, if a method takes just a `Tag[A]` but
it requires a `Tag[Emit[A]]`, the macro will fail and the final complete
tag should be made available instead. The codebase was adapted for this
new restriction.

### Notes

The new implementation addresses several issues but still has some
important limitations. I'm planning to open follow up tickets for the
following:

- **Opaque type support**: The macro isn't able to introspect opaque
types. I've attempted a few times with different APIs but it seems not
possible without relying on compiler internals or `typeSymbol.info`,
which is marked as experimental. Izumi just uses the internal compiler
types directly. It'd be nice to have a proper supported solution so we
can still keep Kyo without any use of internal or experimental language
features. Due to this limitation, opaque types like `Maybe` need to
define the tag manually exposing its internal types, which requires
`Tag.deriveDynamic` for tags using `Maybe`.
- **Fix Async.Join**: I had trouble to make the tag inference work with
the output type as `Result[Nothing, *]` and decided to go with `Any`
instead to unblock this change.
- **Support tags of tags**: The new impl can't infer `Tag[Tag[A]]`,
which is why I used the `TypeMapMarker` in `TypeMap`
- **Compact string representation**: The generated tag string is quite
large now since java serialization isn't very efficient. I've added gzip
compaction to alleviate it a bit but we should introduce a custom
serialization and dictionary-based compaction like in the previous
version.
- **Tag cache customization**: The parameters of the cache are
hardcoded. I'll follow up reviewing all static configs and introducing a
better way to read user-defined configurations since this is popping up
in multiple places of the codebase.
- **Remove erased tags**: Migrate the codebase to use `<:<` where
possible and remove erased tags from effects.

---------

Co-authored-by: Adam Hearn <22334119+hearnadam@users.noreply.github.com>