[data] Introduce Chain + migrate Tag to use it (#1180)
## Problem

We currently don't have a good alternative to using arrays directly.
`Chunk` boxes primitives and requires wrapper objects, which can be
expensive, and `IArray` doesn't offer optimized methods (`getOrElse`
requires a function for example)

## Solution

Introduce a `KArray` in `kyo-data`, which is a wrapper of `Array` with
optimized methods.

## Notes

I've introduced it to get rid of the last allocation in the type checks
of the new tag implementation. After it, type checks don't allocate any
objects, even when the there's a cache miss! The tag impl only allocates
for dynamic tags and when parsing static queries, which happens once for
each static tag.