[data] Chunk optimizations (#1184)
### Problem

The overhead of the default Scala collection methods with `Chunk` is the
main bottleneck in `StreamBench`.

### Solution

- Introduce a thread-local cache in `ChunkBuilder` to reuse buffers
instead of having to allocate and grow them dynamically, which
introduces allocations.
- Introduce an optimized `foreach` similar to `copyTo` allowing
traversal without allocations besides the allocation of the function
parameter
- Override other methods like `map` and `filter` to use the optimized
`foreach`. To avoid having to allocate the `ChunkBuilder` + the function
to call `foreach`, I've introduced `ChunkBuilder.initTransform` that
merges the function implementation in the same allocation as
`ChunkBuilder`.