[core] stable fiber identity (#1190)
### Problem

When a fiber encounters a child fiber/promise to wait on during
execution, `IOTask` currently resumes the parent in a new `IOTask`
instance after the child completes. Besides the performance cost, this
makes the identity of a fiber not stable since it might be executing in
a linked `IOTask` instance if it got suspended by a child fiber. It's a
blocker to implement fiber trace dumps for example since the actual
execution trace might in a linked `IOTask` instance.

### Solution

Adapt the logic to simply resume the `IOTask` after a child completes
and the execution of the parent needs to resume.