[core] Channel/Queue closeAndWaitEmpty (#1191)
Fixes #721

### Problem

`Queue` and `Channel` don't offer a way to close them only after all
pending items are consumed. This is an important limitation in
producer/consumer scenarios where the producer has has finished
processing but it needs to wait for consumers to drain the queue or
channel before closing it.

### Solution

Introduce `Channel.closeAndWaitEmpty` and `Queue.closeAndWaitEmpty`