[core] fix Channel closed exception frame (#1198)
### Problem

We're still getting build
[failures](https://github.com/getkyo/kyo/actions/runs/15153342865/job/42603387092#step:6:6692)
and I noticed the exception is reporting the wrong `Frame`. The initial
frame is kept to report where the closed resource was created but the
actual frame of the exception should come from the current operation
being executed.

### Solution

Properly propagate the `Frame`. This might help identify the issue since
I'm not able to reproduce it locally cc/ @steinybot