[core] move fromCompletionStage from Fiber to Async (#1195)
### Problem

`Fiber` is a low-level effect with complex semantics for newcomers. Its
functionality is essential for integrations and more low-level code but
`Async` provides better usability and covers a majority of use cases.
For this reason, `Fiber`'s API has become more limited over time and we
don't even provide common combinators like `race` for it since the same
functionality can be achieved via `Async` combinators and `Async.run` to
obtain a `Fiber`. I hadn't noticed that `fromCompletionStage` is still
in `Fiber` and using an old method naming pattern with a `Fiber` suffix.

### Solution

Move the method to `Async` and remove `fromCompletionStageFiber`.