[core] channel.pendingPuts/pendingTakes + fix flaky tests (#1203)
### Problem

We have a few flaky channel tests due to the necessity of having pending
puts in some scenarios. Currently, there isn't an easy way to inspect
the number of pending operations, which can be used to fix the
flakiness.

### Solution

Introduce `pendingPuts` and `pendingTakes`. Fix the flaky tests using
them.

### Notes

I'll wait for the build to pass and merge to fix the build. Please feel
free to provide feedback after merge and I'll address it!