[kernel][core] Resource improvements and fix (#1194)
### Problem

When there's a failure, `Resource` is currently not properly waiting for
the finalizers to execute before proceeding. Additionally, it doesn't
provide a way for finalizers to conditionally execute depending on the
outcome of the computation.

### Solution

Fix the handling in case of failure to ensure the computation waits for
finalizers and adapt `Safepoint`, `IO`, and `Resource` to pipe through a
`Maybe[Throwable]` to indicate if the computation failed or not so
finalizers can implement custom logic.