[core] implement `Async.raceFirst` (#1229)
### Problem
`Async.race` completes on the the first success or the last failure. In
certain cases, this is undesirable. For example racing a potentially
failing computation with a promise that may never complete. The
resulting computation will hang indefinitely.

### Solution
Create a new specialized `Race[E, A]` class extending `IOPromise` in
`Fiber`. This will be used to implement customization of Race state
without additional overhead of a separate class. This PR adds support
for `Success` (existing) and `First` (new).