Stream variance (#1268)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

Addresses #679 

### Problem
<!--
Explain here the context, and why you're making this change. What is the
problem you're trying to solve?
-->

To this point `Stream[V, S]` was invariant with `V`. This has led to
some usability issues, especially when merging streams. For instance, in
an application you may have an event type `DomainEvent` with multiple
subtypes, e.g., `UserEvent` and `SystemEvent`. If you had separate
streams of these that eventually need to be merged, you could not simply
call `userEventStream.merge(systemEventStream)`, as these have different
types and could not be lifted by the compiler to `DomainEvent` (or
`UserEvent | SystemEvent` in the case where there is no common
supertype).

### Solution
<!--
Describe your solution. Focus on helping reviewers understand your
technical approach and implementation decisions.
-->

- Now that tags support subtyping and unions, and `Emit[V]` is
contravariant with `V`, we can simply make `Stream[V, S]` covariant with
`V`.
- Since pretty much every method takes `Tag[Emit[Chunkl[V]]]` as an
implicit parameter, however, we've needed to lift `V` to `VV >: V` for
all these methods to satisfy the compiler's variance constraints.
- This change to the Stream methods' type signatures has made it
impossible for the compiler to distinguish between the pure and
effectful overloads of functions like `map`, `filter`, `takeWhile`, etc
(any method that takes a lambda from `V` to some value). Accordingly,
these methods have been split into `map`/`mapKyo`, `filter`/`filterKyo`,
`takeWhile`/`takeWhileKyo`, etc. While it's unfortunate that the
compiler can't distinguish these automatically, I believe it's for the
best since the implementations are substantially different and it's
probably a good thing that the user is forced to be explicit about it.

### Notes
<!--
Add any important additional information as bullet points, such as:
- Implementation details reviewers should know about
- Open questions and concerns
- Limitations
-->

Also added a `dropWhile` method alongside `dropWhileKyo`. I believe it
was an oversight that we didn't have overloads for this one.

---------

Co-authored-by: Adam Hearn <22334119+hearnadam@users.noreply.github.com>