Pipe variance (#1281)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

### Problem
<!--
Explain here the context, and why you're making this change. What is the
problem you're trying to solve?
-->

Composability of `Pipe`, `Sink`, and `Stream` is improved vastly with
variance in their streaming element types. Now that `Sink` and `Stream`
have been updated, `Pipe` is all that's left.

### Solution
<!--
Describe your solution. Focus on helping reviewers understand your
technical approach and implementation decisions.
-->

Now that `Poll[V]` is covariant with `V` and `Emit[V]` is contravariant
with `V`, `Pipe[A, B, S]` can be made contravariant with `A` and
covariant with `B` simply by adding the variance and adding a type
parameter narrowing and widening `A` and `B` respectively.

### Notes
<!--
Add any important additional information as bullet points, such as:
- Implementation details reviewers should know about
- Open questions and concerns
- Limitations
-->

Also added `empty` methods to `Sink` and `Pipe`, and added scaladoc to
all `empty` methods