KyoApp: install signal handler for interruption (#1260)
### Problem
Currently KyoApp doesn't terminate the computation on user signal
(Control+C).

### Solution
Use Async.raceFirst to interrupt an executing computation via a race
with a Promise. The promise is completed with an Interrupt on signal
handle.

### Notes
Signals are not installed on windows as they are not supported as far as I can tell.

/fixes #491