[core] improve IO suspension naming (#1308)
### Problem

Since the renaming of `IO` to `Sync`, its `apply` method doesn't seem as
intuitive anymore since it doesn't indicate suspension of side effects.
Also, `Async.apply` is confusing because it's meant to suspend side
effects but it can give the impression that it'll run the computation
concurrently.

### Solution

Rename the `apply` method of both `Sync` and `Async` to `io` since it's
a common naming to indicate side effects.