[prelude] parametrized generic aspects (#1327)
### Problem

`Aspect`s can't be used in generic contexts when their inputs or outputs
depend on a parametrized type. This issue was raised as part of the work
to generalize `Parse` in https://github.com/getkyo/kyo/pull/1305.

### Solution

Instead of relying on the object identity of the `Aspect` instance,
manage `Cut` activations based on a pair of the type signature of the
`Aspect` via `Tag` and its allocation site via `Frame`.

### Notes

- This approach doesn't keep a strict identity as before since the
`Frame` can be shared in case it's available a scope with multiple
`Aspect` instantiations but that's something that should happen only in
Kyo's codebase since users shouldn't manage `Frame`s.
- I've removed the `default` cut to simplify the implementation since
it's unused.