[kernel] lifted functions fixes (#1325)
### Problem

@ahoy-jon found a couple of issues with lifted functions. Due to the
definition of the implicit conversions being in the same scope as the
definition of `opaque type <`, the compiler doesn't introduce a `lift`
conversion, which creates issues with nested computations in generic
scopes. Additionally, the compiler is having issues with inlining of
pure functions in generic contexts that requires an explicit type
annotation.

### Solution

Introduce the missing `lift` calls in the function conversions and add
explicit type annotations in methods that take functions.