Scoped Fiber init (#1379)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

Fixes #1167 

### Problem
<!--
Explain here the context, and why you're making this change. What is the
problem you're trying to solve?
-->

By default, `Fiber.init` forks a fiber without ever guaranteeing it will
be interrupted. The result is that fibers can easily end up lost in an
application in a permanently suspended or long-running state. It would
be good to have a behavior somewhat like ZIO, that guarantees cleanup of
fibers

### Solution
<!--
Describe your solution. Focus on helping reviewers understand your
technical approach and implementation decisions.
-->

Following the pattern established by #1168 , we change `Fiber.init` to
introduce the `Scope` effect, interrupting the forked fiber in the
registered finalizer; we add a `Fiber.use` method to use a forked fiber
within a function and clean it up afterwards; and finally, we add
`Fiber.initUnscoped` to produce the previous behavior, which forks a
fiber and never guarantees interruptions.

### Notes
<!--
Add any important additional information as bullet points, such as:
- Implementation details reviewers should know about
- Open questions and concerns
- Limitations
-->

While this change was inspired by ZIO's behavior, it departs from ZIO in
that by default ZIO binds each forked fiber to the parent fiber and
interrupts child fiber whenever its parent fiber completes. I chose not
to try to reproduce this for two reasons:
1. It would require some deeper changes to the runtime
2. I don't actually like ZIO's behavior so much, because it means the
behavior of async effects depends on how they are composed. See for
instance:

```scala
import zio.stream._
import zio._

Unsafe.unsafe { implicit unsafe =>
  val eff1 = for
    fiberA <- (ZIO.sleep(1.second) *> Console.printLine("A")).fork
    fiberB <- (ZIO.sleep(1.second) *> Console.printLine("B")).fork
    _ <- Console.printLine("1")
  yield ()

  val eff2 = eff1 *> (ZIO.sleep(5.seconds) *> Console.printLine("2"))

  val eff3 = eff1.fork *> (ZIO.sleep(5.seconds) *> Console.printLine("3"))

  val eff4 = eff1.fork.flatMap(f => (ZIO.sleep(5.seconds) *> f.join *> Console.printLine("4")))

  Runtime.default.unsafe.run(eff1).getOrThrowFiberFailure()
  Runtime.default.unsafe.run(eff2).getOrThrowFiberFailure()
  Runtime.default.unsafe.run(eff3).getOrThrowFiberFailure()
  Runtime.default.unsafe.run(eff4).getOrThrowFiberFailure()
}
```

`eff1` on its own will just print `1`
`eff2` will print `1`, `A`, `B`, `2`
`eff3` will print `1`, `2`
`eff4` will print `1`, `4`

If you construct an effect that depends on ephemeral forked tasks, it
will behave properly in a test or when used in an application in a
forked handler, but when you sequence it other effects it may behave
differently.

In conclusion, I think introducing the notion of parent fiber and child
fiber makes it harder to reason about the fiber lifecycle than if fibers
are just treated the same way as other resources.