[data][prelude] allow bypassing dynamic tag check + support dynamic tags in Aspect.init (#1389)
### Problem

For performance reasons, we don't allow dynamic tags to be inferred
within the `kyo` package. This is a good optimization in general but it
isn't a strict requirement. For example, in
https://github.com/getkyo/kyo/pull/1305 we had to inline methods that
don't make sense to inline to work it around, which isn't ideal.

### Solution

Introduce `Tag.dynamic` to allow bypassing the check within the `kyo`
package and use it in `Aspect.init` to avoid the unnecessary inlining in
https://github.com/getkyo/kyo/pull/1305. It seems reasonable as an
exceptional case.