[kernel] isolate variance for restore + convenience isolate.nest method (#1388)
### Problem

`Isolate`'s new `Restore` type parameter is currently invariant and
prevents referring to an isolate with fewer restored effects as an
isolate with more effects, which is ok given that the pending set is
contravariant. Additionally, we currently don't have a convenient way to
separate a specific effect out of the pending ones.

### Solution

Make `Restore` contravariant and add `isolate.nest`.

### Notes

- I had issues with the compiler inferring `Any` for `Restore`, which is
too restrictive and fails in a few scenarios. I had to add a `Restore <:
Remove` constraint in the `given` to help the compiler.
- I've renamed the second `IsolateTest` in `kyo-prelude` to
`IsolatePreludeTest` to avoid confusion.