Derive ConcreteTag unions/intersections from given component tags (#1397)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

Prerequisite for #1393. See
[discussion](https://github.com/getkyo/kyo/pull/1393#discussion_r2227227541).

### Problem
<!--
Explain here the context, and why you're making this change. What is the
problem you're trying to solve?
-->

The derivation macro for `ConcreteTag` currently does not search for
given `ConcreteTag` instances for components `A` and `B` in composite
types like `A | B` or `A & B`. This forces users to require implicit
evidence for the full composite type in cases where only one of the
types is generic (e.g., `def someMethod[E](A < Abort[E | Closed])(using
ConcreteTag[E | Closed]) ...`). This creates a more severe problem when
one of the types in the intersection isn't known in the type signature,
like when one of the types is a dependent type.

### Solution
<!--
Describe your solution. Focus on helping reviewers understand your
technical approach and implementation decisions.
-->

Add an implicit search in the `createSingle` method of the `ConcreteTag`
derivation macro to look for a given instance prior to trying to
construct it from the `TypeRepr`.

### Notes
<!--
Add any important additional information as bullet points, such as:
- Implementation details reviewers should know about
- Open questions and concerns
- Limitations
-->

Also updated some names which hadn't been updated from the previous name
`SafeClassTag`.